# al-arabi-platform/docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Al-Arabi Ticketing Platform API
  description: |
    Comprehensive API for the Al-Arabi Arabic-first ticketing platform.
    
    ## Features
    - **Multi-language Support**: Arabic and English content
    - **Authentication**: OTP-based phone authentication with JWT tokens
    - **Payment Gateways**: Multiple Arab and international payment providers
    - **Real-time Updates**: WebSocket connections for live data
    - **Geolocation**: PostGIS-powered location services
    - **Contract Management**: PDF generation and e-signature workflow
    
    ## Authentication
    Most endpoints require authentication via JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    API calls are rate limited per user/IP. Current limits:
    - 100 requests per 15 minutes for authenticated users
    - 20 requests per 15 minutes for unauthenticated users
    
    ## Error Handling
    All errors follow the standard format:
    ```json
    {
      "success": false,
      "message": "Error message in English",
      "message_ar": "رسالة الخطأ بالعربية",
      "errors": [...] // Validation errors if applicable
    }
    ```
  version: 1.0.0
  contact:
    name: Al-Arabi API Support
    email: api-support@al-arabi.com
    url: https://docs.al-arabi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://al-arabi.com/terms
servers:
...
paths:
...
  /auth/generate-otp:
    post:
      tags: [Authentication]
      summary: Generate OTP for phone verification
      description: Sends a 6-digit OTP to the specified phone number for various purposes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, purpose]
              properties:
                phone:
                  type: string
                  pattern: '^\+[1-9]\d{1,14}$'
                  example: '+966501234567'
                  description: Phone number in international format
                purpose:
                  type: string
                  enum: [registration, login, password_reset, phone_verification]
                  example: registration
      responses:
        200:
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        409:
          description: Phone number already registered (for registration purpose)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        429:
          $ref: '#/components/responses/RateLimit'
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Register a new user account with OTP verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, otp, first_name, last_name]
              properties:
                phone:
                  type: string
                  pattern: '^\+[1-9]\d{1,14}$'
                  example: '+966501234567'
                otp:
                  type: string
                  pattern: '^\d{6}$'
                  example: '123456'
                first_name:
                  type: string
                  minLength: 2
                  maxLength: 50
                  example: 'أحمد'
                last_name:
                  type: string
                  minLength: 2
                  maxLength: 50
                  example: 'محمد'
                first_name_ar:
                  type: string
                  example: 'أحمد'
                last_name_ar:
                  type: string
                  example: 'محمد'
                email:
                  type: string
                  format: email
                  example: 'ahmed@example.com'
                password:
                  type: string
                  minLength: 8
                  example: 'SecurePass123!'
                preferred_language:
                  type: string
                  enum: [ar, en]
                  default: ar
      responses:
        201:
          description: User registered successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/User'
                          tokens:
                            $ref: '#/components/schemas/TokenPair'
        400:
          $ref: '#/components/responses/BadRequest'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login with OTP
      description: Authenticate user with phone number and OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, otp]
              properties:
                phone:
                  type: string
                  pattern: '^\+[1-9]\d{1,14}$'
                  example: '+966501234567'
                otp:
                  type: string
                  pattern: '^\d{6}$'
                  example: '123456'
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/User'
                          tokens:
                            $ref: '#/components/schemas/TokenPair'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Get a new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
      responses:
        200:
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          access_token:
                            type: string
                            example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
        401:
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user profile
      description: Retrieve the authenticated user's profile information
      security:
        - BearerAuth: []
      responses:
        200:
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/User'
        401:
          $ref: '#/components/responses/Unauthorized'

  # Events Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 'Operation completed successfully'
        message_ar:
          type: string
          example: 'تمت العملية بنجاح'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: 'An error occurred'
        message_ar:
          type: string
          example: 'حدث خطأ'
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        phone:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        first_name_ar:
          type: string
        last_name_ar:
          type: string
        preferred_language:
          type: string
          enum: [ar, en]
        profile_image_url:
          type: string
          format: uri
        is_verified:
          type: boolean
        role:
          type: string
          enum: [user, organizer, admin, super_admin]
        created_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
          example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
        refresh_token:
          type: string
          example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
...
  responses:
    BadRequest:
      description: Bad request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimit:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...